Add chance to suspend/resume the application of hot updates
Keep track of the number of updates while suspended.
When the suspension ends, apply the hot update if there is only one;
otherwise notify that a full reload is needed.
This is necessary for performance reasons because multiple hot updates
can't be aggregated and they have to be applied sequentially.